requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.4
cssselect2==0.7.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cleanup==7.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.3.1
django-postalcodes-mexico==0.6.0
djangorestframework==3.14.0
fonttools==4.39.3
headless-pdfkit==0.1.5
html5lib==1.1
idna==2.10
importlib-metadata==5.1.0
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.6.0
PyJWT==2.6.0
pyphen==0.14.0
python3-openid==3.2.0
pytz==2022.6
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
urllib3==1.26.13
webencodings==0.5.1
zipp==3.11.0
zopfli==0.2.2
